Handle formatting on client
